update health check test for 24a

+++ healthcheck/test_webappserver_refarch_existingvpc.py
@@ -14,7 +14,7 @@ import random
 from datetime import date
 import sys
 
-def main(keypairname, password, ipAddress, SSLCertificateARN, location_arg, platform_arg):
+def main(keypairname, password, ipAddress, location_arg, platform_arg):
    # Reference architectures in production.
    # Deploy a stack for creating VPC with 2 subnets
     existing_template_url = "https://matlab-web-app-server-templates.s3.amazonaws.com/r2022a_refarch/VPCStack.yml"
@@ -32,46 +32,34 @@ def main(keypairname, password, ipAddress, SSLCertificateARN, location_arg, plat
                   {'ParameterKey': 'AdminIPAddress', 'ParameterValue': ipAddress},
                   {'ParameterKey': 'Password', 'ParameterValue': password},
                   {'ParameterKey': 'ConfirmPassword', 'ParameterValue': password},
-                  {'ParameterKey': 'SSLCertificateARN', 'ParameterValue': SSLCertificateARN},
                   {'ParameterKey': 'WorkerSystem', 'ParameterValue': platform_arg},
-                  {'ParameterKey': 'WorkerInstanceType', 'ParameterValue': 'm5.xlarge'},
                   {'ParameterKey': 'UseSameIPForClient', 'ParameterValue': 'Yes'},
                   {'ParameterKey': 'ExistingVPC', 'ParameterValue': vpc_id},
-                  {'ParameterKey': 'ExistingVPCAddress', 'ParameterValue': vpc_cidr},
                   {'ParameterKey': 'ExistingSubnet1', 'ParameterValue': subnet1},
                   {'ParameterKey': 'ExistingSubnet2', 'ParameterValue': subnet2},
-                  {'ParameterKey': 'ExistingSubnet3', 'ParameterValue': subnet3},
-                  {'ParameterKey': 'DeployLicenseServer', 'ParameterValue': 'Yes'},
-                  {'ParameterKey': 'ExistingLicenseServer', 'ParameterValue': '0.0.0.0'}]
+                  {'ParameterKey': 'DeployLicenseServer', 'ParameterValue': 'Yes'}]
 
     # Find latest MATLAB release from Github page and get template url text
     res = requests.get(f"https://github.com/mathworks-ref-arch/{ref_arch_name}/blob/master/releases/")
 
     # Deploy a stack for the latest two releases
     latest_releases = [re.findall("releases/(R\d{4}[ab]\\b)", res.text)[-1], re.findall("releases/(R\d{4}[ab]\\b)", res.text)[-2]]
-    number_of_releases = 2
+    number_of_releases = 1
     for i in range(number_of_releases):
         matlab_release = latest_releases[i]
         print("Testing Health Check Release: " + matlab_release + "\n")
         github_base_dir = "https://raw.githubusercontent.com/mathworks-ref-arch"
 
-        # Getting template url from .txt file
-        # NOTE for code cleanup: Delete this branch after next release
-        if matlab_release == "R2021b":
-            template_url_path = f"{github_base_dir}/{ref_arch_name}/master/releases/{matlab_release}/templates.txt"
-            file = urllib.request.urlopen(template_url_path)
-            template_url = file.readline().decode("utf-8").rstrip()
-
         # Getting template url from .json file
-        else:
-            template_url_path = f"{github_base_dir}/{ref_arch_name}/main/releases/{matlab_release}/templates.json"
-            response = urllib.request.urlopen(template_url_path)
-            template_json = json.loads(response.read())
-            template_url = template_json["WebAppServer_existing.yml"]
+        template_url_path = f"{github_base_dir}/{ref_arch_name}/main/releases/{matlab_release}/templates.json"
+        response = urllib.request.urlopen(template_url_path)
+        template_json = json.loads(response.read())
+        template_url = template_json["WebAppServer_existing.yml"]
+
         stack_name = "mwas-refarch-health-check-existing-vpc" + matlab_release + date.today().strftime('%m-%d-%Y') + str(random.randint(1,101))
         ct = datetime.datetime.now()
         print("Date time before deployment of stack:-", ct)
-
+        stack=None
         # Deploying the stack
         try:
             print("deploying the stack")
@@ -80,13 +68,14 @@ def main(keypairname, password, ipAddress, SSLCertificateARN, location_arg, plat
         except Exception as e:
             raise (e)
         finally:
-           # Delete the deployment
-           print("deleting the stack")
-           deploy.delete_stack(stack)
-           print("success deleting the stack")
-           ct = datetime.datetime.now()
-           print("Date time after deployment and deletion of stack:-", ct)
-           print("\n\n")
+           if stack:
+            # Delete the deployment
+            print("deleting the stack")
+            deploy.delete_stack(stack)
+            print("success deleting the stack")
+            ct = datetime.datetime.now()
+            print("Date time after deployment and deletion of stack:-", ct)
+            print("\n\n")
 
 if __name__ == '__main__':
-    main(sys.argv[1], sys.argv[2], sys.argv[3], sys.argv[4], sys.argv[5], sys.argv[6])
+    main(sys.argv[1], sys.argv[2], sys.argv[3], sys.argv[4], sys.argv[5])
